backend/requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pandas
python-dotenv
fastapi
uvicorn
sqlalchemy
cryptography